[DEV-5731] Make parse_request_repositories() generic
